+++ src/main/java/duke/Storage.java
@@ -1,9 +1,16 @@
+package duke;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
 import java.util.List;
+import duke.task.Task;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Todo;
+
 
 public class Storage {
     protected Path filePath;
