+++ src/main/java/duke/TaskList.java
@@ -1,4 +1,7 @@
+package duke;
+
 import java.util.ArrayList;
+import duke.task.Task;
 
 public class TaskList {
     protected ArrayList<Task> list;
