+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 public class ExitCommand extends Command {
     public static final String INSTRUCTION = "bye";
