+++ src/main/java/duke/command/IncorrectCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 public class IncorrectCommand extends Command {
     String errorMessage;
