+++ text-ui-test/sources.txt
@@ -1,0 +1,6 @@
+../src/main/java/duke/Duke.java
+../src/main/java/duke/DukeException.java
+../src/main/java/duke/Parser.java
+../src/main/java/duke/Storage.java
+../src/main/java/duke/TaskList.java
+../src/main/java/duke/Ui.java
